Amalgamate simple setter slash commands into /configure
